+++ src/strings.js
@@ -6,14 +6,15 @@
 
 
 /**
- * Convert an {Uint8Array} with utf8 data to string
+ * Convert an {Uint8Array} to string
+ * Can be used after {os.read}
  *
- * @param {Uint8Array} array
+ * @param {Uint8Array} array to convert
  * @param {integer} len array length (optional)
  *
  * @return {string}
  */
-const utf8ArrayToStr = (array, len) => {
+const bytesArrayToStr = (array, len) => {
     if (undefined === len) {
         len = array.length;
     }
@@ -50,6 +51,50 @@ const utf8ArrayToStr = (array, len) => {
     }
 
     return finalStr;
+}
+
+/**
+ * Convert a {string} to a {Uint8Array}
+ * Can be use with {os.write}
+ *
+ * Borrowed from https://gist.github.com/joni/3760795
+ *
+ * @param {string} str string to convert
+ *
+ * @return {Uint8Array}
+ */
+const strToBytesArray = (str) => {
+    const arr = [];
+    let charCode;
+
+    for (let i = 0, length = str.length; i < length; ++i) {
+        charCode = str.charCodeAt(i);
+        if (charCode < 0x80) {
+            arr.push(charCode);
+        }
+        else if (charCode < 0x800) {
+            arr.push(0xc0 | (charCode >> 6),
+                     0x80 | (charCode & 0x3f));
+        }
+        else if (charCode < 0xd800 || charCode >= 0xe000) {
+            arr.push(0xe0 | (charCode >> 12),
+                     0x80 | ((charCode>>6) & 0x3f),
+                     0x80 | (charCode & 0x3f));
+        }
+        // surrogate pair
+        else {
+            ++i;
+            // UTF-16 encodes 0x10000-0x10FFFF by
+            // subtracting 0x10000 and splitting the
+            // 20 bits of 0x0-0xFFFFF into two halves
+            charCode = 0x10000 + (((charCode & 0x3ff)<<10) | (str.charCodeAt(i) & 0x3ff));
+            arr.push(0xf0 | (charCode >>18),
+                      0x80 | ((charCode>>12) & 0x3f),
+                      0x80 | ((charCode>>6) & 0x3f),
+                      0x80 | (charCode & 0x3f));
+        }
+    }
+    return Uint8Array.from(arr);
 }
 
 /**
@@ -97,6 +142,7 @@ const getLines = (content, incompleteLine, skipBlankLines = false) => {
 }
 
 export {
-    utf8ArrayToStr,
+    bytesArrayToStr,
+    strToBytesArray,
     getLines
 }